# Don't even run `self.check_field_type()` in `AST.__init__` unless type checks are enabled