This method has been run >500,000 times in the `get_movie()` benchmark query
alone.  At this rate micro-optimizations like the one employed provide actual
time savings. In fact, it's kinda stupid but it's a 5% speedup.

Before:
```
compile_edgeql_script: Mean +- std dev: 2.62 ms +- 0.10 ms
compile_edgeql_script_webapp_configure_system: Mean +- std dev: 43.2 ms +- 3.5 ms
compile_edgeql_script_webapp_get_movie: Mean +- std dev: 250 ms +- 67 ms
compile_edgeql_script_webapp_get_person: Mean +- std dev: 295 ms +- 88 ms
compile_edgeql_script_webapp_get_user: Mean +- std dev: 125 ms +- 7 ms
compile_edgeql_script_webapp_insert_movie: Mean +- std dev: 161 ms +- 4 ms
compile_edgeql_script_webapp_insert_person: Mean +- std dev: 62.6 ms +- 1.4 ms
compile_edgeql_script_webapp_insert_review: Mean +- std dev: 93.8 ms +- 2.3 ms
```

After:
```
compile_edgeql_script: Mean +- std dev: 2.44 ms +- 0.07 ms
compile_edgeql_script_webapp_configure_system: Mean +- std dev: 40.1 ms +- 1.3 ms
compile_edgeql_script_webapp_get_movie: Mean +- std dev: 245 ms +- 67 ms
compile_edgeql_script_webapp_get_person: Mean +- std dev: 281 ms +- 88 ms
compile_edgeql_script_webapp_get_user: Mean +- std dev: 118 ms +- 5 ms
compile_edgeql_script_webapp_insert_movie: Mean +- std dev: 153 ms +- 4 ms
compile_edgeql_script_webapp_insert_person: Mean +- std dev: 58.6 ms +- 1.7 ms
compile_edgeql_script_webapp_insert_review: Mean +- std dev: 88.9 ms +- 2.3 ms
```